write Demo_v0-3-1

+++ meal-planner/Meals/IMeal.cs
@@ -6,6 +6,6 @@ namespace meal_planner.Meals
     public interface IMeal
     {
         IRecipe Recipe();
-        int Servings();
+        double Servings();
     }
 }
